deployer: ensure desired tag is in ISes
technical fix for https://github.com/openshift/origin/issues/8948
In order to ensure the the triggers on the DCs fire, seed the tags on
each ImageStream with the desired version. Otherwise, if there are
more than 5 tags available, the IS may not have the desired tag
imported automatically (or may have it imported by another name).